backend/requirements.txt
# Core Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# Web Scraping & Data Processing
requests==2.31.0
beautifulsoup4==4.12.2

# Geocoding
geopy==2.4.0